Cleanup and add serialization of ranges
This commit makes some of the code a bit cleaner and readable, replacing 's',
'so' with 'start', 'startOffset', etc.

In addition, and more importantly, the annotator now adds annotations
(including /most/ multirange annotations) to its internal registry, with
ranges stored in a format that can be easily serialized to a database and
redrawn on the page. Deserialization is not currently supported.

Removed redundant xpath filtering code.