Make JSON serialization/deserialization work.
This commit adds a jQuery JSON parsing/generating plugin, for the purposes of
demo.html only -- not a requirement for the installed jsannotate. Fixed up the
deserialization code and added a basic UI in demo.html to dump and load
annotations.

Annotator#register is now a little more clever, and will try and work out what
kind of range its being asked to register.