Refactor Annotator into jQuery plugin.
This commit adds support for using the annotator as a jQuery plugin, i.e. it
is now instantiated on a specific element:

    $(elem).annotator();

Method calls on the annotator can be performed too:

    $(elem).annotator('myMethod', arg1, arg2);

although to preserve jQuery's chaining, any return values will be lost unless
you use an alternative syntax:

    $(elem).data('annotator').myMethod(arg1, arg2);

All this (including a new means of event delegation in DelegatorClass) is
intended to facilitate the creation of a separate AnnotationStore plugin which
will listen to events raised on the element by Annotator, and will persist
annotations through a REST server interface.

NB: a few specs are broken at the moment: problems with the specs, not the
code, but these will be fixed soon.