Annotator can now de serialise annotation in XML documents. Fixes #54
Previously the annotator failed to parse the XPaths stored on the
annotations when loaded into an XHTML document served as
application/xhtml+xml. We now check to see if the current document is an
XML one and if it is try to resolve any namespaced elements first by using
a namespace resolver created with document.createNSResolver() then falling
back to a cruder method of defining a custom resolver and appending a
custom namespace to all nodes.

Also added a dev.xhtml file for specifically testing the injection of
annotations into an XHTML document. This can be removed once we have a
working browser test runner.