Error is no longer thrown in Firefox when hiding Viewer. Fixes #64
This issue was caused by Firefox passing an additional "lateness" parameter
to setTimeout() callbacks. We now check the existence of the event object
as well as the .preventDefault() method before calling. This is done using
a new util.preventEventDefault() helper method.

Also removed the unnecessary $.proxy() wrapper around the setTimeout()
callback as the method is bound to the viewer at runtime anyway.