Add ability to set default permissions on the basis of an authToken.
This adds the ability for the Permissions plugin to automatically set up a sensible default set
of permissions, and to detect the current user, when the Auth plugin is loaded.

I've also upgraded the plugin-sniffing mechanism used by the Store plugin to be inline with
that used elsewhere in the code, namely looking at @annotator.plugins.PluginName.