Prevent Annotator from interfering with selections fully outside its element
At the moment, if you make a selection that is fully outside the annotated
element, Annotator will remove the selection instantly. This isn't very
helpful if you're trying to copy some text without annotating it.

This change ensures that these selections aren't tampered with by Annotator.

Question for the future: currently, selections that overlap with the annotator
wrapper are truncated to the beginning (or end) of the wrapper. Perhaps
instead we should *split* the selection (of a single range) into two ranges,
which both remain in the document, but only one of which is annotated?