refactor selector into a plugin
With this change, selection utilities become a plugin making it
easier to override the default behavior in order to provide new
selection possibilities, such as multimedia selections.

More work needs to be done if the core needs to be completely ignorant
of the format of selections, but this is a start.

Also reduces the number of places tracking selection and mouse event
state in order to properly record the selection for a new annotation.