Use empty string if prefix is unset or null
This fixes #175. Previously, the prefix defaulted to '/', which results in
`_urlFor` returning URLs of the form '//annotations', which the browser
interprets as a protocol-agnostic full URL.

The fallback prefix should be '', and the check for existence of
`@options.prefix` should not magically convert the empty string into a single
slash.