Provide emulation of JSON post in Store plugin
As requested in #180 (see also #172), this adds an `emulateJSON` option to the
Store plugin, which allows Annotator to communicate with legacy webservers
that don't much like POSTS with `application/json` type payloads. Instead, the
data to be sent is JSON encoded client-side and is sent as a standard form
request (with content-type `application/x-www-form-urlencoded`) with the JSON
stored in the `json` key.

In addition, if `emulateHTTP` is also set, the Store will add a `_method` key
to the form, containing the same value as the `X-HTTP-Method-Override` header.