don't choke when an annotation has no highlights
With this change, 'annotationDeleted' events are fired for
cancelled annotations. This is a behavior change. Previously,
this event was only fired when saved annotations were deleted,
but not when new edits were cancelled.

It's my belief that this event is helpful information. A quick
audit of Annotator shows that nothing relies upon the previous
behavior. For example, the Store plugin registers annotations
after the 'annotationCreated' event is fired and ignores events
on unregistered annotations, thereby ignoring the 'annotationDeleted'
event that this change causes.