Remove redo build system in favour of a Makefile
The redo build system, while simple and elegant once you got to know it,
confused a lot of people. redo is not well known, and even though "do" was
shipped in the tools/ directory, it still resulted in cognitive overhead that we
shouldn't be demanding of our users.

You can now build the Annotator (and bookmarklet) by typing simply

    make