registry: Add a Registry object for annotations
In order to make custom Store plugins feasible, a lot of the work currently
performed by the Store plugin needs to move back into Annotator itself. The
Registry object is the part of Annotator that will look after communication with
a simplified Store plugin. It will have responsibility for sanity checking store
responses, and ensuring a common interface between other parts of Annotator and
the store.