Be more conservative when splitting text elements
Some browsers have the habit of splitting off 0-length text nodes,
when they are told to do so. (Thus littering the DOM with useless elements.)
Therefore, let's try to be more conservative when slicing and dicing elements.

(This modification only adds one extra condition,
and a bunch of explanatory comments.)