Util.mousePosition - use a proper positioning root
This partially reverts d45a36662cf904134783091931a8dd85e7ebaf2a by
restoring the use of the jQuery `.offset()` method instead of
`.position()` but fixes the function for the case where the passed
in element (usually the .annotator-wrapper div) is not a positioning
root. In this case, its offset parent is taken as the offset element.

Fix #243
Fix #244