Clean up Viewer and plugin tests
Many plugins expected DOM events to be used for the transport of (e.g.)
'beforeAnnotationCreated' events. This was fixed to use Delegator custom
events.

A number of plugins instantiated an Annotator with a default Store
plugin. These have been updated to use the NullStore.

Various other fixes.

And a passing test suite :)