Add module-based plugin support
With this it should be easy for authors of all kinds to build on Annotator.

Creates index.js, main.js, and plugin.js in the pkg directory.

The index.js module provides a way for packagers of downstream applications
to require annotator easily in a CommonJS environment. For bundling Annotator
into a single script application, browserify can be used straightforwardly.
For creating plugins with browserify, ./lib/annotator can be externalized.
See the description of plugin.js for details.

The main.js module eases integration with AMD environments such as require.js.
It resolves to the UMD wrapped build output.

The plugin.js module serves as the basis for packaging plugins with browserify.
It populates the Annotator.Plugin namespace as a side effect of requiring the
'__plugin' module and re-exports Annotator. In this way, it can be used as the
entry point for creating a UMD module which will augment the Annotator
through the side-effects of including the plugin.

  $ browserify -r ./node_modules/annotator/plugin \
        -r ./node_modules/annotator:annotator \
        -r ./tags:__plugin \
        --noparse=annotator \
        --standalone=Annotator \
        > annotator.tags.js

The remaining thing to do is to build the .coffee to pkg/lib during release so
that the CommonJS require of annotator can resolve the full source if desired,
such as when making a downstream application that is meant to be distributed
as a single bundle.