Export plugin packaging tools
Now it should be easy to build plugins in separate repositories.

The basic trick for plugins is to develop them assuming a module
system.

Plugins developed as AMD first will be portable, but may have to be
unwrapped with deamdify for browserify consumers. CommonJS plugins
should be usable by both communities without trouble.

The node package of annotator now exports a browserify factory that
helps to standalone releases of plugins for consumers that choose
not to use a module system. The built in plugins are packaged this
way at build time to be compatible with the old release structure.

The browser field and the main.js it points to should ensure that
AMD and CommonJS authors can easily depend on Annotator.

In a separate commit, I'll add a build step for releasing ./lib so
that these systems can find all the internals.