Fix namespace reference error
Try this:

Make a new project. Then `npm install annotator`.

  `browserify -r annotator -x jquery > bundle.js`

That gives you annotator locally. Your application can depend on it.
Using require('annotator') or setting up your node_modules as a package
in r.js should work in both AMD and CommonJS environments.

  `browserify ./:annotator/plugin/example -r annotator:annotator -x jquery

This command would build index.js and bundle annotator, exposing a global
require function for 'annotator' and 'annotator/plugin/example'.

From node, the annotator module has a utility function which takes an
instance of browserify and any options to pass to the factory. Then, it
requires and externalizes annotator in such a way that any further exposed
modules are automatically required by an injected loader entry point. This
gives module authors an easy way to bundle UMD modules that augment the global
Annotator when scripts are concatenated. It is used by the build system
to generate the plugin bundles, but it can be used to mangle the namespace
in any way or monkey-patch Annotator.

Hopefully, this will help the community publish small modules rather than
requiring consensus on changes to the annotator core and in time many pieces
can be factored out or adapted to changing interface requirements and diverse
environments.