Extract AnnotationProvider, replacing NullStore
Instead of the registry dealing directly with annotations, the registry is
merely a runtime coordination object for bootstrapping applications that use
the annotator and include optional submodules. The registry is now the thinnest
I can imagine making it.

The bulk of what was in the registry is moved to src/annotations.coffee, which
provides the core integration with storage, but is kept separate. This makes it
possible to load a different provider for the registry, such as a Backbone
model that implements the AnnotationProvider interface, or any other such
thing that might make observing the annotations more idiomatic for the rest of
the application.