Remove redundant build writes
I can't recall exactly what I was doing here, but I think there was
an iteration of the browserify branch where I was attempting to write
all the plugins out as annotator-full was bundled. This is not needed
and rather confusing. The plugin builds take care of themselves. With
the inclusion of dependencies from node_modules, like the standalone
backbone components, this was causing those files to be written over
and the changed mtime was causing annotator and annotator-full to
appear out of date to make when the other was built.