Remove redundant Make targets and includes
The dependency generation step in the main pkg/%.js rule has a sed
line that ensures the generated .d files are structured as rules
such that each only applies to the source files it strictly depends
on. As such, there's no need for a more complicated include at the
end than $(DEPDIR)/*.d to catch everything.

Also, the bootstrap script is caught by this because the bookmarklet
target depends on pkg/bootstrap.js.