Restore support for annotationData in the Store
Recent changes to the Store plugin broke support for the annotationData
option (fdbfd6ad3f2f50191079bdd4fafcce27de136b8f).

This commit restores the test case and the attendant code in the Store
plugin, but deprecates the use of this option in favour of integrators
simply using the `beforeAnnotationCreated` and `beforeAnnotationUpdated`
hooks to achieve the same behaviour in a more explicit way.