Bump version to 2.0.0-dev.2
The make dist result works for an npm publish. We should update
the make_release tool to just `npm publish` the resulting tarball.
It seems npm does not care about the name of the top directory in
the tarball, as long as it contains the package.json.

This is better than npm pack because that will cause the top
directory to be "package", which is less friendly if we decide
to distribute the tarball directly via github releases.