Move the NullStore plugin into the plugins directory
The only reason I didn't do this before is because explicitly
registering the plugin inside `nullstore.coffee` wouldn't work, due to a
circular dependency between `annotator.coffee` and `nullstore.coffee`.

This commit gets around this problem by introducing the idea of a "core
plugin", one that is built and shipped inside the core build of
Annotator, and one that is registered in `annotator.coffee`, not the
plugin source file. The new UI plugins (extractions of the current UI)
will take this form.