[WIP] Remove the wrapper element
I've realised while doing this refactoring that the only reason for
needing the wrapper element was as an offset parent for positioning UI
elements.

There's really no harm in just adding UI elements to the end of
document.body, at which point all we need to do is calculate the offset
of document.body and use that when positioning UI elements.

This commit removes the wrapper (and temporarily disables the viewer and
editor, which still depend on it) and updates the Adder and Highlights
plugins.