Generalizing adder button: data format
When we detect a selection, we store some data, which is later
used when the adder button is clicked.

Before this change, we stored the list of ranges, and then when the
adder button was clicked, we created an annotation out of it.

To make this more generic, I moved the code responsible for creating
the annotation from the list of ranges from the adder button click
to the earlier phase (after detecting a selection).

So now the skeleton for the annotation is created right away, when
a selection is detected, and when the adder button is clicked,
we simply use this skeleton to create the real animation.

(This way, we can decouple the analysis of the selection from the adder
button.)