Use wicked-good-xpath library instead of xpath.js
The originally used xpath.js was not up-to-date and had problems when
executing the provided document.evaluate().

Wicked-good-xpath is a non-invasive up-to-date fast solution.

Changelog
annotator.coffee: removed one bit ugly hack to get the xpath.js library
instead of that: If wgxpath is present, just install it.
Installing the wgxpath does not hurt browsers which already have the
xpath functionality.

*.html: added wgxpath.install as a dependency instead of xpath.
Please note that IE 11+ no longer checks for conditional comments.
(So <!--[IE]> is no longer working) And also it is not needed because
wgxpath install is not invasive.

With this all the tests are successful under IE11 and both
the demo.html and dev.html are working too.