textselector: Don't use NormalizedRange#toRange()
The `toRange()` method of `NormalizedRange` is the only point of
coupling between the DOM implementation of the W3C Range APIs, and the
xpath-range library. By inlining the tiny amount of code need to convert
a `NormalizedRange` into a DOM Range in the text selector, we can remove
this coupling.

This was motivated by openannotation/xpath-range#4.