core: Extract Annotator core functionality
Here's another attempt (or start of one) to improve on the plugin
architecture of Annotator. I've stripped away a lot of complexity,
including:

- Factory (gone completely)
- Annotator.Plugin.register (Annotator.Plugin is just a plain old object
  again)

But the most important change is the addition of src/core.coffee, which
contains AnnotatorCore. AnnotatorCore is:

- where plugins and storage implementations are registered
- responsible for propagating hook actions ('onAnnotationCreated') to
  plugin instances
- able to tear down an Annotator instance

That's all at the moment, but that leaves the Annotator class itself to
be responsible *only* for setting up a "default Annotator application"
-- the thing that most users have been using throughout the 1.2.x
series.

I've commented out a bunch of plugin tests -- the ones that still rely
on nasty direct inter-plugin dependencies.