ui: Extract Editor component and plugin
Just as with the Highlighter in a13e72d, this commit moves the bulk of
the Editor functionality into the UI module. There remains a skeletal
Editor plugin, which admittedly isn't terribly useful on its own, but it
is included for completeness.

As part of this work, the #load() method of the Editor was refactored to
return a promise (resolving if the editor is submitted, rejecting if
cancelled) rather than having a "private" event handler on the Editor
itself.