ui/filter: Remove Delegator dependency
Delegator is really not doing an awful lot here, and if anything hinders
readability. I've replaced the use of Delegator with direct calls to
jQuery's event delegation method (".on()"), using an event namespace to
simplify event unbinding.

While it looks like I'm cheating by removing the tests for this event
binding, these kind of tests are ridiculous anyway -- they're testing
the implementation of the Filter plugin, and not its unit behaviour.
They won't be missed.