Move namespace to standalone annotator-plugintools package
Currently, plugins in the core distribution (as well as hypothetical
3rd-party plugins) can

    require('annotator')

and get a reference to Annotator via the magic of src/namespace.js,
which inspects the current environment. If we're in a browser, and the
global `Annotator` is available, then we use that. In particular, this
requires that when we browserify the plugin bundles, we "trick" them
into thinking that the 'annotator' name refers to src/namespace.js.

This commit makes this process less magical, by introducing a standalone
package, "annotator-plugintools". This contains the shim previously
known as src/namespace.js.