Replace chai with assertive-chai
IE8 can't deal with the use of ES5 in the chai library. Luckily, the
relevant part of chai is used only by the expect/should assertion
format, which we don't use. assertive-chai includes only the assert API
as a wrapper around Node's assert package, which has a corresponding
browserify shim.