Specify a single node version
Rather bizarrely, adding an explicit matrix include (for running the
lint tools) exposes a Travis bug in which it attempts to call

    nvm use ["0.10"]

and everything blows up.

Seeing as we don't care about matrix builds for multiple versions of
node, we can just pin this to the latest stable version of node.