Fix horrible issue with karma-bro
We still (after close to 2 days digging into this) don't understand
what's going wrong here. Despite apparently having restored all package
versions to ones that previously worked, we've been unable to get this
working again.

While debugging the issue we discovered that adding a bundle call in the
prebundle hook gets the tests passing again. This is a horrible fix and
doesn't relieve us of the need to understand what's going on here and
how to fix it, but I don't want to burn any more time on this problem
for now.