Fix and simplify build process
Translating Annotator back to plain old JavaScript makes it far easier
for users to use the uncompressed CommonJS source. We can simply push
the repository as-is to NPM (excluding what we wish with .npmignore),
allowing others to use browserify, jspm, or other solutions as they
wish.

We still want to provide prepackaged minified files for those users who
aren't familiar with this world, and this commit fixes the build scripts
to bring back those compiled assets.