Reorganise module layout
This commit renames the major modules and packages that comprise
Annotator for consistency with the wider JavaScript ecosystem. Modules
now are referenced with lowercase names,

    Util -> util
    Storage -> storage

and so on.

In addition, one important change is that Annotator now exports the root
module to the page, rather than the Annotator constructor. Once
included, the appropriate invocation is now

    new annotator.Annotator(...)

rather than the old

    new Annotator(...)