'start' hook receives app, not registry
Now we have a clear distinction between configuration and startup
(namely, the 'configure' and 'start' hooks) it makes more sense for the
app instance, and not the registry, to be passed into the 'start' hook.

This obviates the need to bind properties (such as "annotations", or
"notify") to both the registry and the application, leaving the registry
solely used for configuration.