Remove tags/markdown from ui.main by default, add extensions
In v1.2.x the tags and markdown plugins aren't on by default, so they
probably shouldn't be on by default in the v2.0 user interface.

This commit removes these from the default UI and adds options to
annotator.ui.main which allow for configuration of arbitrary extensions
of Editor and Viewer, as well as the Viewer renderer.