viewer: don't repeatedly show overlapping annotations
See GH issue #520.

If you have a chunk of text with hundreds of annotations at the same
spot, the markup looks like this:

```html
<span class="annotation-hl">
  <span class="annotator-hl">
    ...
      Foo
    ...
  </span>
</span>
```

Mousing over the text fires this event once per `<span>`; so `.load()`
would be called once per `<span>`, each time with the full set of
annotations. `.load()` is surprisingly expensive --
`Widget.checkOrientation` seems costly -- and so in my test case with
150 annotations on the same block of text, the browser would freeze for
~10 seconds.

(I tried changing the highlighter to use a single `<span>` for all
annotations covering a piece of text, but it's surprisingly fiddly to
get right if you have annotations which intersect but are not
identical.)