Drop the extensions altogether
These are actually way too problematic as they stand and it's holding
up progress. In the alpha.3 tag, these were built as standalone bundles
that depended upon annotator to be global using browserify-shim. There
is actually no sane way to make these portable across environments that
doesn't involve re-instating the namespace module / plugin tool. An
alternative is to re-think how these extension modules are defined, such
as to require invoking them as functions with `annotator` passed explictly
so that the application developer can provide the proper annotator.

In the meantime, this is really holding up 2.0 and these don't seem like
they should be so critical as to block release. If there is demand, these
should be recreated as separate repositories and some further effort put into
finalizing a packaging story for them.